pyproject.toml
[project]
name = "llm-finance-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.18.4 ; python_full_version >= '3.10'",
    "annotated-doc==0.0.4 ; python_full_version >= '3.8'",
    "annotated-types==0.7.0 ; python_full_version >= '3.8'",
    "anyio==4.12.1 ; python_full_version >= '3.9'",
    "attrs==25.4.0 ; python_full_version >= '3.9'",
    "audioop-lts==0.2.2 ; python_full_version >= '3.13'",
    "certifi==2026.2.25 ; python_full_version >= '3.7'",
    "cffi==2.0.0 ; python_full_version >= '3.9' and platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.5 ; python_full_version >= '3.7'",
    "click==8.3.1 ; python_full_version >= '3.10'",
    "cryptography==46.0.5",
    "distro==1.9.0 ; python_full_version >= '3.6'",
    "et-xmlfile==2.0.0 ; python_full_version >= '3.8'",
    "fastapi==0.135.1 ; python_full_version >= '3.10'",
    "greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "h11==0.16.0 ; python_full_version >= '3.8'",
    "httpcore==1.0.9 ; python_full_version >= '3.8'",
    "httptools==0.7.1",
    "httpx==0.28.1 ; python_full_version >= '3.8'",
    "httpx-sse==0.4.3 ; python_full_version >= '3.9'",
    "idna==3.11 ; python_full_version >= '3.8'",
    "jiter==0.13.0 ; python_full_version >= '3.9'",
    "jsonpatch==1.33 ; python_full_version >= '2.7'",
    "jsonpointer==3.0.0 ; python_full_version >= '3.7'",
    "jsonschema==4.26.0 ; python_full_version >= '3.10'",
    "jsonschema-specifications==2025.9.1 ; python_full_version >= '3.9'",
    "langchain-core==1.2.17 ; python_full_version >= '3.10' and python_full_version < '4'",
    "langgraph==1.0.10 ; python_full_version >= '3.10'",
    "langgraph-checkpoint==4.0.1 ; python_full_version >= '3.10'",
    "langgraph-prebuilt==1.0.8 ; python_full_version >= '3.10'",
    "langgraph-sdk==0.3.9 ; python_full_version >= '3.10'",
    "langsmith==0.7.14 ; python_full_version >= '3.10'",
    "mako==1.3.10 ; python_full_version >= '3.8'",
    "markupsafe==3.0.3 ; python_full_version >= '3.9'",
    "mcp==1.26.0 ; python_full_version >= '3.10'",
    "numpy==2.4.2 ; python_full_version < '3.14'",
    "openai==2.26.0 ; python_full_version >= '3.9'",
    "openpyxl==3.1.5 ; python_full_version >= '3.8'",
    "orjson==3.11.7 ; python_full_version >= '3.10'",
    "ormsgpack==1.12.2 ; python_full_version >= '3.10'",
    "packaging==26.0 ; python_full_version >= '3.8'",
    "pandas==3.0.1 ; python_full_version >= '3.11'",
    "psycopg2-binary==2.9.11 ; python_full_version >= '3.9'",
    "pycparser==3.0 ; implementation_name != 'PyPy'",
    "pydantic==2.12.5 ; python_full_version >= '3.9'",
    "pydantic-core==2.41.5 ; python_full_version >= '3.9'",
    "pydantic-settings==2.13.1 ; python_full_version >= '3.10'",
    "pyjwt[crypto]==2.11.0 ; python_full_version >= '3.9'",
    "pypdf2==3.0.1 ; python_full_version >= '3.6'",
    "python-dateutil==2.9.0.post0 ; python_full_version >= '2.7'",
    "python-dotenv==1.2.2 ; python_full_version >= '3.10'",
    "python-multipart==0.0.22 ; python_full_version >= '3.10'",
    "pyyaml==6.0.3",
    "referencing==0.37.0 ; python_full_version >= '3.10'",
    "requests==2.32.5 ; python_full_version >= '3.9'",
    "requests-toolbelt==1.0.0 ; python_full_version >= '2.7'",
    "rpds-py==0.30.0 ; python_full_version >= '3.10'",
    "six==1.17.0 ; python_full_version >= '2.7'",
    "sniffio==1.3.1 ; python_full_version >= '3.7'",
    "speechrecognition==3.14.6 ; python_full_version >= '3.9'",
    "sqlalchemy==2.0.48 ; python_full_version >= '3.7'",
    "sse-starlette==3.3.2 ; python_full_version >= '3.10'",
    "standard-aifc==3.13.0 ; python_full_version >= '3.13'",
    "standard-chunk==3.13.0 ; python_full_version >= '3.13'",
    "starlette==0.52.1 ; python_full_version >= '3.10'",
    "tenacity==9.1.4 ; python_full_version >= '3.10'",
    "tqdm==4.67.3 ; python_full_version >= '3.7'",
    "typing-extensions==4.15.0 ; python_full_version >= '3.9'",
    "typing-inspection==0.4.2 ; python_full_version >= '3.9'",
    "urllib3==2.6.3 ; python_full_version >= '3.9'",
    "uuid-utils==0.14.1 ; python_full_version >= '3.9'",
    "uvicorn[standard]==0.41.0 ; python_full_version >= '3.10'",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0 ; python_full_version >= '3.10'",
    "xxhash==3.6.0 ; python_full_version >= '3.7'",
    "zstandard==0.25.0 ; python_full_version >= '3.9'",
]
